Changed names of models and new dissipation model for mergers.
-names of models were all modified to reflect the paper from which they
were taken (thus avoiding generic names such as GALFORM, l-galaxies,
etc.).

- new dissipation model included ala Hopkins et al. (2009). This model
modified the sizes of merger remnants by taking into account the gas
fraction of the mergers.

+++ include/numerical_constants.h
@@ -77,9 +77,9 @@ namespace constants {
 	constexpr float GIGA =1.0e+9, logGIGA =+9.0;
 
 	/* Small numbers (for tolerances usually).*/
-	constexpr float EPS3=1.0e-3;
-	constexpr float EPS4=1.0e-4;
-	constexpr float EPS6=1.0e-6;
+	constexpr double EPS3=1.0e-3;
+	constexpr double EPS4=1.0e-4;
+	constexpr double EPS6=1.0e-6;
 
 	/*Unit conversions.*/
 	constexpr float J2ERG=1.0e7, logJ2ERG=7.0; /*Number of ergs in a Joule.*/
